+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch project states
 - DataSHIELD status
 - Filter draft projects to project manager admin
+- If type is DataSHIELD, use template opal-ccp and output format Opal